# Use more straightforward compression approach and buffer pool