--- a/apm-lambda-extension/extension/apm_server.go
+++ b/apm-lambda-extension/extension/apm_server.go
@@ -35,54 +35,43 @@ var bufferPool = sync.Pool{New: func() interface{} {
 //       connection open across invocations?
 func PostToApmServer(agentData AgentData, config *extensionConfig) error {
 	endpointURI := "intake/v2/events"
-	var request *http.Request
+	encoding := agentData.ContentEncoding
+	buf := bufferPool.Get().(*bytes.Buffer)
+	defer func() {
+		buf.Reset()
+		bufferPool.Put(buf)
+	}()
 
 	if agentData.ContentEncoding == "" {
-		buf := bufferPool.Get().(*bytes.Buffer)
-		defer func() {
-			buf.Reset()
-			bufferPool.Put(buf)
-		}()
-
+		encoding = "gzip"
 		gw, err := gzip.NewWriterLevel(buf, gzip.BestSpeed)
 		if err != nil {
 			return err
 		}
-
 		if _, err := gw.Write(agentData.Data); err != nil {
 			log.Printf("Failed to compress data: %v", err)
 		}
 		if err := gw.Close(); err != nil {
 			log.Printf("Failed write compressed data to buffer: %v", err)
 		}
-
-		req, err := http.NewRequest("POST", config.apmServerUrl+endpointURI, buf)
-		if err != nil {
-			return fmt.Errorf("failed to create a new request when posting to APM server: %v", err)
-		}
-		req.Header.Add("Content-Encoding", "gzip")
-		request = req
 	} else {
-		buf := bufferPool.Get().(*bytes.Buffer)
 		buf.Write(agentData.Data)
-		req, err := http.NewRequest("POST", config.apmServerUrl+endpointURI, buf)
-		if err != nil {
-			return fmt.Errorf("failed to create a new request when posting to APM server: %v", err)
-		}
-		req.Header.Add("Content-Encoding", agentData.ContentEncoding)
-		request = req
 	}
 
-	request.Header.Add("Content-Type", "application/x-ndjson")
+	req, err := http.NewRequest("POST", config.apmServerUrl+endpointURI, buf)
+	if err != nil {
+		return fmt.Errorf("failed to create a new request when posting to APM server: %v", err)
+	}
+	req.Header.Add("Content-Encoding", encoding)
+	req.Header.Add("Content-Type", "application/x-ndjson")
 	if config.apmServerApiKey != "" {
-		request.Header.Add("Authorization", "ApiKey "+config.apmServerApiKey)
+		req.Header.Add("Authorization", "ApiKey "+config.apmServerApiKey)
 	} else if config.apmServerSecretToken != "" {
-		request.Header.Add("Authorization", "Bearer "+config.apmServerSecretToken)
+		req.Header.Add("Authorization", "Bearer "+config.apmServerSecretToken)
 	}
 
 	client := &http.Client{}
-
-	resp, err := client.Do(request)
+	resp, err := client.Do(req)
 	if err != nil {
 		return fmt.Errorf("failed to post to APM server: %v", err)
 	}