Instead of using `io.Copy`, use a more straightforward compression approach. Also use a buffer pool to avoid creating a new buffer each time a request is sent.