@felixbarny 2021-11-17
Is there a specific issue with `io.Pipe`? I think the advantage of it is that there are no buffers required at all.