@estolfo 2021-11-17
Yes, there are no synchronization guarantees and the client could start issuing the request before the gzip writer had finished writing everything. If there's an error, the body would be incomplete. Using a buffer ensures that the data is properly gzipped before it is sent.