@felixbarny 2021-11-17
> the client could start issuing the request before the gzip writer had finished writing everything. If there's an error, the body would be incomplete.

I don't see an issue with that. In fact, I think that's preferable. It avoids unnecessarily copying from buffer to buffer. The Elastic APM Java agent is also directly serializing the events to the HTTP request body to avoid having to copy and allocate buffers. I have never seen a situation where compression would fail.