@marclop 2021-11-18
I modified the PostToAPM function to accept an `*http.Client` and it doesn't cause an allocation every time it is called and [did some benchmarking on my machine](https://gist.github.com/marclop/6fb7021e5ccb5b1350024655298479e0) and the pipes seem to be slower than the pooled `bytes.Buffer` approach most likely because the `io.Pipe()` call causes extra allocations.

Overall it seems to be around 10% faster than the `io.Pipe` approach and it causes less allocations, which is also desirable.

```
$ benchstat old.txt new.txt
name         old time/op    new time/op    delta
PostToAPM-8     151µs ±11%     135µs ± 8%  -10.88%  (p=0.032 n=5+5)

name         old alloc/op   new alloc/op   delta
PostToAPM-8    1.25MB ± 0%    1.22MB ± 0%   -2.46%  (p=0.008 n=5+5)

name         old allocs/op  new allocs/op  delta
PostToAPM-8       112 ± 1%       104 ± 1%   -7.33%  (p=0.008 n=5+5)
```